# openapi/power-alert.v1.yaml
openapi: 3.1.0
info:
  title: Power Alert API
  version: 1.0.0
  description: Privileged command API for Power Alert 2.0.
servers:
  - url: https://api.poweralert.example/v1
security:
  - firebaseBearer: []
    appCheck: []
paths:
  /health:
    get:
      security: []
      summary: Liveness and dependency status
      responses:
        "200":
          description: Healthy
  /outages:
    post:
      summary: Create an outage
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutage"
      responses:
        "201":
          description: Outage created
        "400":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
  /outages/{outageId}/status:
    patch:
      summary: Advance or correct outage status
      parameters:
        - $ref: "#/components/parameters/OutageId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, message]
              properties:
                status:
                  $ref: "#/components/schemas/OutageStatus"
                message:
                  type: string
                  minLength: 3
                  maxLength: 500
                estimatedRestoreTime:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Status updated
  /complaints:
    post:
      summary: Create a complaint after uploads complete
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaint"
      responses:
        "201":
          description: Complaint created
  /complaints/{complaintId}/assign:
    post:
      summary: Assign a technician
      parameters:
        - name: complaintId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [technicianId]
              properties:
                technicianId: { type: string, minLength: 1 }
                note: { type: string, maxLength: 500 }
      responses:
        "200":
          description: Complaint assigned
  /complaints/{complaintId}/status:
    patch:
      summary: Update complaint workflow status
      parameters:
        - name: complaintId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, note]
              properties:
                status:
                  type: string
                  enum: [submitted, verified, repairStarted, restoring, resolved, rejected, cancelled]
                note: { type: string, minLength: 3, maxLength: 500 }
                repairPhotoPaths:
                  type: array
                  maxItems: 8
                  items: { type: string }
      responses:
        "200":
          description: Complaint updated
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    appCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    OutageId:
      name: outageId
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            type: object
            required: [type, title, status, code, requestId]
            properties:
              type: { type: string, format: uri }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }
              code: { type: string }
              requestId: { type: string }
  schemas:
    OutageStatus:
      type: string
      enum: [reported, investigating, identified, repairing, restoring, restored, scheduled, emergency, cancelled]
    CreateOutage:
      type: object
      additionalProperties: false
      required: [areaIds, reason, severity, priority, startTime, affectedPopulation]
      properties:
        providerId:
          type: string
          description: Required only when a super admin acts for a provider.
        areaIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string }
        reason: { type: string, minLength: 3, maxLength: 300 }
        severity: { type: string, enum: [low, medium, high, critical] }
        priority: { type: string, enum: [normal, urgent, emergency] }
        startTime: { type: string, format: date-time }
        estimatedRestoreTime: { type: string, format: date-time }
        affectedPopulation: { type: integer, minimum: 0 }
        feederId: { type: string }
        transformerId: { type: string }
    CreateComplaint:
      type: object
      additionalProperties: false
      required: [areaId, issueType, description, latitude, longitude]
      properties:
        areaId: { type: string }
        issueType:
          type: string
          enum: [noPower, lowVoltage, transformerFault, poleDamage, wireDamage, streetLight, meterIssue, other]
        description: { type: string, minLength: 10, maxLength: 1000 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        attachmentPaths:
          type: array
          maxItems: 8
          items: { type: string }